refactor(narrative): remove expandable-itin; use default-itin
