+++ lib/components/narrative/narrative-itinerary.js
@@ -22,12 +22,9 @@ export default class NarrativeItinerary extends Component {
     if (onClick) {
       onClick()
     } else if (!active) {
-      console.log('setting to', index)
-      setActiveItinerary({index})
+      setActiveItinerary && setActiveItinerary({index})
     } else {
-      // FIXME
-      console.log('setting to null')
-      setActiveItinerary({index: null})
+      setActiveItinerary && setActiveItinerary({index: null})
     }
   }
