+++ lib/components/narrative/tabbed-itineraries.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react'
 import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
-import { setActiveItinerary, setActiveLeg, setActiveStep, setUseRealtimeResponse } from '../../actions/narrative'
+import * as narrativeActions from '../../actions/narrative'
 import DefaultItinerary from './default/default-itinerary'
 import { getActiveSearch, getRealtimeEffects } from '../../util/state'
 
@@ -39,8 +39,10 @@ class TabbedItineraries extends Component {
       itineraries,
       itineraryClass,
       realtimeEffects,
+      setActiveItinerary,
+      timeFormat,
       useRealtime,
-      timeFormat
+      ...itineraryClassProps
     } = this.props
     if (!itineraries) return null
 
@@ -55,56 +57,14 @@ class TabbedItineraries extends Component {
       <div className='options itinerary tabbed-itineraries'>
         <div className='tab-row'>
           {itineraries.map((itinerary, index) => {
-            const timeOptions = {
-              format: timeFormat,
-              offset: getTimeZoneOffset(itinerary)
-            }
-            const classNames = ['tab-button', 'clear-button-formatting']
-            const { caloriesBurned } = calculatePhysicalActivity(itinerary)
-            const {
-              centsToString,
-              maxTNCFare,
-              minTNCFare,
-              transitFare
-            } = calculateFares(itinerary)
-            // TODO: support non-USD
-            let minTotalFare = minTNCFare * 100 + transitFare
-            const plus = maxTNCFare && maxTNCFare > minTNCFare ? '+' : ''
-            if (index === activeItinerary) classNames.push('selected')
             return (
-              <Button
-                key={`tab-button-${index}`}
-                className={classNames.join(' ')}
-                // FIXME
-                onClick={() => { this.props.setActiveItinerary({index}) }}
-              >
-                <div className='title'><span>Option {index + 1}</span></div>
-                <div className='details'>
-                  {/* The itinerary duration in hrs/mins */}
-                  {formatDuration(itinerary.duration)}
-
-                  {/* The duration as a time range */}
-                  <span>
-                    <br />
-                    {formatTime(itinerary.startTime, timeOptions)} - {formatTime(itinerary.endTime, timeOptions)}
-                  </span>
-
-                  {/* the fare / calories summary line */}
-                  <span>
-                    <br />
-                    {minTotalFare ? <span>{`${centsToString(minTotalFare)}${plus}`} &bull; </span> : ''}
-                    {Math.round(caloriesBurned)} Cal
-                  </span>
-
-                  {/* The 'X tranfers' line, if applicable */}
-                  {itinerary.transfers > 0 && (
-                    <span>
-                      <br />
-                      {itinerary.transfers} transfer{itinerary.transfers > 1 ? 's' : ''}
-                    </span>
-                  )}
-                </div>
-              </Button>
+              <TabButton
+                index={index}
+                isActive={index === activeItinerary}
+                itinerary={itinerary}
+                onClick={setActiveItinerary}
+                timeFormat={timeFormat}
+              />
             )
           })}
         </div>
@@ -122,9 +82,10 @@ class TabbedItineraries extends Component {
             index: activeItinerary,
             key: activeItinerary,
             active: true,
+            expanded: true,
             routingType: 'ITINERARY',
             showRealtimeAnnotation,
-            ...this.props
+            ...itineraryClassProps
           })
           : null
         }
@@ -156,9 +117,75 @@ const mapStateToProps = (state, ownProps) => {
   }
 }
 
+class TabButton extends Component {
+  _onClick = () => {
+    const {index, onClick} = this.props
+    // FIXME: change signature once actions resolved with otp-ui
+    onClick(index)
+  }
+
+  render () {
+    const {index, isActive, itinerary, timeFormat} = this.props
+    const timeOptions = {
+      format: timeFormat,
+      offset: getTimeZoneOffset(itinerary)
+    }
+    const classNames = ['tab-button', 'clear-button-formatting']
+    const { caloriesBurned } = calculatePhysicalActivity(itinerary)
+    const {
+      centsToString,
+      maxTNCFare,
+      minTNCFare,
+      transitFare
+    } = calculateFares(itinerary)
+    // TODO: support non-USD
+    const minTotalFare = minTNCFare * 100 + transitFare
+    const plus = maxTNCFare && maxTNCFare > minTNCFare ? '+' : ''
+    if (isActive) classNames.push('selected')
+    return (
+      <Button
+        key={`tab-button-${index}`}
+        className={classNames.join(' ')}
+        onClick={this._onClick}
+      >
+        <div className='title'><span>Option {index + 1}</span></div>
+        <div className='details'>
+          {/* The itinerary duration in hrs/mins */}
+          {formatDuration(itinerary.duration)}
+
+          {/* The duration as a time range */}
+          <span>
+            <br />
+            {formatTime(itinerary.startTime, timeOptions)} - {formatTime(itinerary.endTime, timeOptions)}
+          </span>
+
+          {/* the fare / calories summary line */}
+          <span>
+            <br />
+            {minTotalFare ? <span>{`${centsToString(minTotalFare)}${plus}`} &bull; </span> : ''}
+            {Math.round(caloriesBurned)} Cal
+          </span>
+
+          {/* The 'X tranfers' line, if applicable */}
+          {itinerary.transfers > 0 && (
+            <span>
+              <br />
+              {itinerary.transfers} transfer{itinerary.transfers > 1 ? 's' : ''}
+            </span>
+          )}
+        </div>
+      </Button>
+    )
+  }
+}
+
 const mapDispatchToProps = (dispatch, ownProps) => {
+  const {setActiveItinerary, setActiveLeg, setActiveStep, setUseRealtimeResponse} = narrativeActions
   return {
-    setActiveItinerary,
+    // FIXME
+    setActiveItinerary: (index) => {
+      dispatch(setActiveItinerary({index}))
+    },
     // FIXME
     setActiveLeg: (index, leg) => {
       dispatch(setActiveLeg({index, leg}))
@@ -167,7 +194,10 @@ const mapDispatchToProps = (dispatch, ownProps) => {
     setActiveStep: (index, step) => {
       dispatch(setActiveStep({index, step}))
     },
-    setUseRealtimeResponse
+    // FIXME
+    setUseRealtimeResponse: ({useRealtime}) => {
+      dispatch(setUseRealtimeResponse({useRealtime}))
+    }
   }
 }
